+++ .settings/org.eclipse.core.resources.prefs
@@ -49,17 +49,13 @@ encoding//plugin.video.vstream/resources/sites/free_telechargement_org.py=utf-8
 encoding//plugin.video.vstream/resources/sites/freebox.py=utf-8
 encoding//plugin.video.vstream/resources/sites/french_stream_lol.py=utf-8
 encoding//plugin.video.vstream/resources/sites/gum_gum_streaming_com.py=utf-8
-encoding//plugin.video.vstream/resources/sites/hds_fm.py=utf-8
 encoding//plugin.video.vstream/resources/sites/ianime.py=utf-8
 encoding//plugin.video.vstream/resources/sites/kepliz_com.py=utf-8
 encoding//plugin.video.vstream/resources/sites/livetv.py=utf-8
-encoding//plugin.video.vstream/resources/sites/lsdb.py=utf-8
-encoding//plugin.video.vstream/resources/sites/mesfilms.py=utf-8
 encoding//plugin.video.vstream/resources/sites/pastebin.py=utf-8
 encoding//plugin.video.vstream/resources/sites/planet_streaming.py=utf-8
 encoding//plugin.video.vstream/resources/sites/siteonefichier.py=utf-8
 encoding//plugin.video.vstream/resources/sites/siteuptobox.py=utf-8
-encoding//plugin.video.vstream/resources/sites/streaminz.py=utf-8
 encoding//plugin.video.vstream/resources/sites/streamonsport.py=utf-8
 encoding//plugin.video.vstream/resources/sites/themoviedb_org.py=utf-8
 encoding//plugin.video.vstream/resources/sites/topimdb.py=utf-8
