+++ plugin.video.vstream/resources/sites/trash/k_streaming.py
