+++ plugin.video.vstream/resources/sites/trash/series_en_streaming.py
