+++ plugin.video.vstream/resources/sites/trash/streaming_series.py
